backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1

sqlalchemy==2.0.28
pymysql==1.1.0

python-multipart==0.0.9
python-dotenv==1.0.1

boto3==1.34.82
requests==2.31.0

torch==2.2.2
torchvision==0.17.2
pillow==10.3.0

scikit-learn==1.4.2
numpy==1.26.4

httpx==0.27.0
pydantic==2.5.2
